requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
astroid==2.15.6
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
dill==0.3.7
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
outcome==1.2.0
packaging==23.1
parameterized==0.9.0
platformdirs==3.10.0
pluggy==1.3.0
pycparser==2.21
PyJWT==2.8.0
pylint==2.17.5
PySocks==1.7.1
pytest==7.4.2
python-dotenv==1.0.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tomlkit==0.12.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
webdriver-manager==4.0.0
wrapt==1.15.0
wsproto==1.2.0
